feat: add ability to get scope info from position
Inspired from TextMate grammar, it constructs a DAG
for different scope present in XML Grammar.
It then matches current line against the RegEx
provided by TextMate Grammar. If matched then
it will store the scope related information in memory.
This information is useful to check triggers for
elements and attributes